+++ src/engine/ohlcv.py
@@ -1,5 +1,4 @@
 from datetime import datetime
-from decimal import Decimal
 from engine.enums import Timeframe
 
 
@@ -8,11 +7,11 @@ class OHLCV:
         self,
         symbol: str,
         timestamp: datetime,
-        open: Decimal,
-        high: Decimal,
-        low: Decimal,
-        close: Decimal,
-        volume: int,
+        open: float,
+        high: float,
+        low: float,
+        close: float,
+        volume: float,
         timeframe: Timeframe,
     ):
         self.symbol = symbol
